Revert "chore(ci): move ExecRunner interface to exec package (#11147)"
This reverts commit a7745d9fa9e50253631cc4ca03f396ae25ad5fc6.
