+++ .ci/magician/cmd/interfaces.go
@@ -39,3 +39,15 @@ type CloudbuildClient interface {
 	ApproveCommunityChecker(prNumber, commitSha string) error
 	TriggerMMPresubmitRuns(commitSha string, substitutions map[string]string) error
 }
+
+type ExecRunner interface {
+	GetCWD() string
+	Copy(src, dest string) error
+	Mkdir(path string) error
+	RemoveAll(path string) error
+	PushDir(path string) error
+	PopDir() error
+	WriteFile(name, data string) error
+	Run(name string, args []string, env map[string]string) (string, error)
+	MustRun(name string, args []string, env map[string]string) string
+}
