+++ .ci/magician/cmd/mock_runner_test.go
@@ -20,7 +20,6 @@ import (
 	"errors"
 	"fmt"
 	"log"
-	"magician/exec"
 	"os"
 	"path/filepath"
 	"sort"
@@ -32,7 +31,7 @@ import (
 type ParameterList []any
 
 type MockRunner interface {
-	exec.ExecRunner
+	ExecRunner
 	Calls(method string) ([]ParameterList, bool)
 }
